docs: create Linux ABI compatibility guide and update documentation with implementation findings (TEAM_217)
- Add `docs/specs/LINUX_ABI_GUIDE.md` documenting critical ABI alignment knowledge (Stat layout, auxv, TLS context switching)
- Update `docs/ARCHITECTURE.md` to link new guide and userspace ABI spec
- Update `docs/specs/userspace-abi.md` to v1.1.0 with corrected Stat structure, syscall status, and auxv specification
- Implement TLS (`TPIDR_EL0`) context switching in `cpu_switch_to` assembly

+++ userspace/ulib/src/fs.rs
@@ -54,7 +54,7 @@ impl File {
             return Err(Error::from_errno(ret));
         }
         Ok(Metadata {
-            size: stat.st_size,
+            size: stat.st_size as u64,
             mode: stat.st_mode,
         })
     }
